added sliders labels and text box for testing

+++ src/ttsBase.java
@@ -4,7 +4,7 @@ import com.sun.speech.freetts.VoiceManager;
 public class ttsBase {
     private static Voice voice;
 
-    public static void tts(String inputText,int rate,int pitch,int volume){
+    public static void tts(String inputText,int rate,int pitch,float volume){
         System.setProperty("freetts.voices", "com.sun.speech.freetts.en.us.cmu_us_kal.KevinVoiceDirectory");
         //Voice voice;//Creating object of Voice class
         voice = VoiceManager.getInstance().getVoice("kevin");//Getting voice
